requirements.txt
gspread==5.10.0
oauth2client==4.1.3
openai>=1.30.0
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.65.0
loguru==0.7.0
tenacity==8.2.2
rapidfuzz
pyyaml==6.0.1
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0